requirements.txt
absl-py==0.7.1
astor==0.7.1
backcall==0.1.0
certifi==2019.6.16
decorator==4.3.0
emoji==0.5.2
gast==0.2.0
glob3==0.0.1
google-pasta==0.1.7
grpcio==1.16.0
h5py==2.8.0
inflect==2.1.0
ipython==7.1.1
ipython-genutils==0.2.0
jedi==0.13.1
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
numpy==1.15.4
pandas==0.23.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.7
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.2.0
python-dateutil==2.7.5
pytz==2018.7
scipy==1.1.0
six==1.11.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
Theano==1.0.3
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.11.2
